Addressing comments from Nikita
