+++ tests/cnf/ran/internal/ranhelper/ranhelper.go
@@ -1,6 +1,10 @@
 package ranhelper
 
 import (
+	"context"
+	"os/exec"
+	"time"
+
 	"github.com/golang/glog"
 	"github.com/openshift-kni/eco-goinfra/pkg/clients"
 	"github.com/openshift-kni/eco-goinfra/pkg/pod"
@@ -52,4 +56,18 @@ func UnmarshalRaw[T any](raw []byte) (*T, error) {
 	}
 
 	return &typed, nil
+}
+
+// ExecLocalCommand runs the provided command with the provided args locally, cancelling execution if it exceeds
+// timeout.
+func ExecLocalCommand(timeout time.Duration, command string, args ...string) (string, error) {
+	glog.V(90).Infof("Locally executing command '%s' with args '%v'", command, args)
+
+	ctx, cancel := context.WithTimeout(context.TODO(), timeout)
+
+	defer cancel()
+
+	output, err := exec.CommandContext(ctx, command, args...).Output()
+
+	return string(output), err
 }
