+++ tests/internal/config/default.yaml
@@ -11,7 +11,7 @@ control_plane_label: "control-plane"
 tc_prefix: "TC-"
 mco_namespace: "openshift-machine-config-operator"
 mco_config_daemon_name: "machine-config-daemon"
-logging_namespace: "openshift-logging"
+logging_operator_namespace: "openshift-logging"
 sriov_operator_namespace: "openshift-sriov-network-operator"
 nmstate_operator_namespace: "openshift-nmstate"
 sriov_fec_operator_namespace: "vran-acceleration-operators"
