+++ tests/internal/params/vars.go
@@ -1,11 +1,9 @@
 package params
 
 // PrivilegedNSLabels represents privileged labels.
-var (
-	PrivilegedNSLabels = map[string]string{
-		"pod-security.kubernetes.io/audit":               "privileged",
-		"pod-security.kubernetes.io/enforce":             "privileged",
-		"pod-security.kubernetes.io/warn":                "privileged",
-		"security.openshift.io/scc.podSecurityLabelSync": "false",
-	}
-)
+var PrivilegedNSLabels = map[string]string{
+	"pod-security.kubernetes.io/audit":               "privileged",
+	"pod-security.kubernetes.io/enforce":             "privileged",
+	"pod-security.kubernetes.io/warn":                "privileged",
+	"security.openshift.io/scc.podSecurityLabelSync": "false",
+}
